Upload MongoDB database
A zip file that you’ll need to unzip and restore using the mongorestore
command. The database is called travelDiary, and contains two
collections: (1) gpsPoints, where each GPS point recorded by the app,
along with the ground truth and all the other information, such as
accelerometer reading and Wi-Fi network name, are stored as an
individual document; and (2) segments, where each trip and activity, as
inferred by our algorithms, is stored as an individual document
(currently, the collection only contains activities, but at some stage
we should also add the trips).